% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MPhStimepoints.R
\name{plot.MPhStimepoints}
\alias{plot.MPhStimepoints}
\title{Plot data of class MPhStimepoints}
\usage{
\method{plot}{MPhStimepoints}(x, title = NULL, ylim = c(-0.015, 0.015), ncol = 1, ...)
}
\arguments{
\item{x}{object of class MPhStimepoints}

\item{title}{character, plot title}

\item{ylim}{numeric vector, a vector with two numeric values, specifying the lower and upper limits of the scale}

\item{ncol}{numerical, number of rows for the sequence of panels}

\item{...}{other graphical parameters}
}
\value{
A \code{hclustering} object.

If \code{k} is \code{NULL}, the \code{hclustering} object is a list of 3 elements:
}
\description{
Plot data of class MPhStimepoints
}
\details{
The \code{MPhSscores} function performs a preliminary standardization of columns in \code{data}.
}
\references{
G.B. Tornielli, M. Sandri, M. Fasoli, A. Amato, M. Pezzotti, P. Zuccolotto and S. Zenoni.
A molecular phenology scale of grape berry development.
Horticulture Research, 2023, 10: uhad048
}
\seealso{
\code{\link[stats]{hclust}}
}
\author{
Marco Sandri, Paola Zuccolotto (\email{sandri.marco@gmail.com})
}
