% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPhSscores.R
\name{MPhSscores}
\alias{MPhSscores}
\title{Project data onto the MPhS subspace}
\usage{
MPhSscores(data, scaling_type = "scale")
}
\arguments{
\item{data}{numeric data frame.}

\item{scaling_type}{character, type of scaling of gene expressions. Available options: \code{none}, \code{scale} (default), \code{means_SDs}.}
}
\value{
A \code{hclustering} object.

If \code{k} is \code{NULL}, the \code{hclustering} object is a list of 3 elements:
}
\description{
Project data onto the MPhS subspace
}
\details{
The \code{MPhSscores} function performs a preliminary standardization of columns in \code{data}.
}
\references{
G.B. Tornielli, M. Sandri, M. Fasoli, A. Amato, M. Pezzotti, P. Zuccolotto and S. Zenoni.
A molecular phenology scale of grape berry development.
Horticulture Research, 2023, 10: uhad048
}
\seealso{
\code{\link[stats]{hclust}}
}
\author{
Marco Sandri, Paola Zuccolotto (\email{sandri.marco@gmail.com})
}
