% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPhStimepoints.R
\name{MPhStimepoints}
\alias{MPhStimepoints}
\title{Project data onto the MPhS subspace}
\usage{
MPhStimepoints(
  data,
  strata_var,
  stage_var,
  scaling_type = "scale",
  gene_keyword = "VIT_"
)
}
\arguments{
\item{data}{numeric data frame.}

\item{strata_var}{vector of character, the name of stratification variable(s)}

\item{stage_var, }{character, the name of variable defining development (stage)}

\item{scaling_type}{character, type of scaling of gene expressions. Available options: \code{none}, \code{scale} (default), \code{means_SDs}.}

\item{gene_keyword}{character, a keyword used to identify all the columns of \code{data} that refers to gene expression (default \code{"VIT_"})}
}
\value{
A \code{hclustering} object.

If \code{k} is \code{NULL}, the \code{hclustering} object is a list of 3 elements:
}
\description{
Project data onto the MPhS subspace
}
\details{
The \code{MPhSscores} function performs a preliminary standardization of columns in \code{data}.
}
\references{
G.B. Tornielli, M. Sandri, M. Fasoli, A. Amato, M. Pezzotti, P. Zuccolotto and S. Zenoni.
A molecular phenology scale of grape berry development.
Horticulture Research, 2023, 10: uhad048
}
\seealso{
\code{\link[stats]{hclust}}
}
\author{
Marco Sandri, Paola Zuccolotto (\email{sandri.marco@gmail.com})
}
